refactor: group acp and acpx agent modules
