+++ packages/browseros-agent/apps/server/src/api/services/agents/agent-harness-service.ts
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: AGPL-3.0-or-later
  */
 
-import { AcpxRuntime } from '../../../lib/agents/acpx-runtime'
+import { AcpxRuntime } from '../../../lib/agents/acpx/runtime'
 import {
   type ActiveTurnInfo,
   type TurnFrame,
