+++ packages/browseros-agent/apps/server/src/lib/agents/runtime/claude-host-process-runtime.ts
@@ -9,12 +9,12 @@ import { logger } from '../../logger'
 import type {
   PrepareAcpxAgentContextInput,
   PreparedAcpxAgentContext,
-} from '../acpx-agent-adapter'
+} from '../acpx/agent-adapter'
 import {
   finishBrowserosManagedContext,
   prepareBrowserosManagedContext,
-} from '../acpx-agent-common'
-import { resolveAgentRuntimePaths } from '../acpx-runtime-context'
+} from '../acpx/agent-common'
+import { resolveAgentRuntimePaths } from '../acpx/runtime-context'
 import { HostProcessAgentRuntime } from './host-process-agent-runtime'
 import { getAgentRuntimeRegistry } from './registry'
 import type { RuntimeDescriptor } from './types'
